% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snake_loader_function.R
\name{snake_loader}
\alias{snake_loader}
\title{Add the Snake Game to a Shiny App!}
\usage{
snake_loader(outputId, header = "Play Snake while you wait!",
  controls = TRUE)
}
\arguments{
\item{outputId}{The output id of the element that triggers the snake game loader;
probably some output that takes a long time to load (or else there won't be time to play!)}

\item{header}{The title to display over the game screen.}

\item{controls}{Should text showing telling user to use the arrow keys appear below the game screen?}
}
\value{
A conditional panel that will load the snake game when the provided \code{outputId} is \code{recalculating}.
}
\description{
Use the snake game to pass time during long running shiny operations.
}
\examples{
\dontrun{
library(shiny)
library(snakeLoadR)

shinyApp(
  shinyUI(
    fluidPage(
      fluidRow(
        column(width=10, offset=1, algin="left",
               actionButton("my_button", "Start Fake 30 Second Job"),
               uiOutput("my_output"),
        )
      ),
      snakeLoadR::snake_loader(outputId = "my_output",
                               header = "Play Snake while you wait!",
                               controls = TRUE)
    )
  ),
  shinyServer(function(input, output) {
    output$my_output <- renderUI({
      if(input$my_button != 0) Sys.sleep(30)
      HTML(paste0("<h3>Fake job completed <code>", input$my_button,"</code> times!</h3>"))
    })
  })
)
}
}
